requirements.txt
django==1.5.4
django-social-auth==0.7.28
